Added authentication for messages ensuring the user belongs to the messages channel and has appropriate permissions.

+++ chat-app-server/src/main/java/com/example/chatappserver/controller/MessagesController.java
@@ -1,28 +1,40 @@
 package com.example.chatappserver.controller;
 
+import com.example.chatappserver.model.CustomUserDetails;
 import com.example.chatappserver.model.Message;
 import com.example.chatappserver.model.Attachment;
 import com.example.chatappserver.model.MessageAttachmentResponse;
 import com.example.chatappserver.repository.MessagesDao;
+import com.example.chatappserver.service.AuthService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.sql.Timestamp;
 import java.util.List;
 
 @RestController
-@RequestMapping("/messages")
+@RequestMapping("/servers/{serverID}/channels/{channelID}/messages")
 public class MessagesController {
     private final MessagesDao messagesDao;
+    private final AuthService authService;
 
     @Autowired
-    public MessagesController(MessagesDao messagesDao) { this.messagesDao = messagesDao; }
+    public MessagesController(MessagesDao messagesDao, AuthService authService) {
+        this.messagesDao = messagesDao;
+        this.authService = authService;
+    }
 
     // Create a message with no attachment
     @PostMapping
-    public ResponseEntity<Integer> createMessage(@RequestBody Message message) {
+    public ResponseEntity<Integer> createMessage(
+            @RequestBody Message message, @PathVariable int channelID,
+            @AuthenticationPrincipal CustomUserDetails user) {
+        if (!authService.userInChannel(user.getUserId(), channelID)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         messagesDao.create(message);
         return ResponseEntity.status(HttpStatus.CREATED).body(message.getMessageID());
     }
@@ -30,7 +42,11 @@ public class MessagesController {
     // Create a message with 1-3 attachments
     @PostMapping("/attachments")
     public ResponseEntity<MessageAttachmentResponse> createMessageWithAttachments(
-            @RequestBody Message message) {
+            @RequestBody Message message, @PathVariable int channelID,
+            @AuthenticationPrincipal CustomUserDetails user) {
+        if (!authService.userInChannel(user.getUserId(), channelID)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         messagesDao.createWithAttachments(message);
         List<Integer> attachmentIds = message.getAttachments().stream()
                 .map(Attachment::getAttachmentID).toList();
@@ -42,9 +58,13 @@ public class MessagesController {
     }
 
     // Get all messages in a Channel
-    @GetMapping("/{channelID}")
+    @GetMapping
     public ResponseEntity<List<Message>> getMessagesInChannel(
-            @PathVariable int channelID) {
+            @PathVariable int channelID,
+            @AuthenticationPrincipal CustomUserDetails user) {
+        if (!authService.userInChannel(user.getUserId(), channelID)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         List<Message> res = messagesDao.getMessagesInChannel(channelID);
         return ResponseEntity.ok(res);
     }
@@ -70,7 +90,13 @@ public class MessagesController {
     }
     // Edit a message text, no change to attachments
     @PutMapping("/{messageID}")
-    public ResponseEntity<Void> editMessage(@RequestBody MessageEditRequest editMessage, @PathVariable int messageID) {
+    public ResponseEntity<Void> editMessage(
+            @RequestBody MessageEditRequest editMessage, @PathVariable int messageID,
+            @AuthenticationPrincipal CustomUserDetails user) {
+        // Checks for the message, ensuring the message exists with the messageID and userID
+        if (!messagesDao.doesMessageExist(messageID, user.getUserId())) {
+            return ResponseEntity.notFound().build();
+        }
         messagesDao.editMessageText(messageID, editMessage.getText(), editMessage.getTime());
         return ResponseEntity.ok().build();
     }
@@ -78,14 +104,27 @@ public class MessagesController {
     // Edit a message by removing attachment(s)
     @PutMapping("/attachments/{messageID}")
     public ResponseEntity<Void> removeAttachment(
-            @PathVariable int messageID, @RequestBody List<Integer> attachmentIDs) {
+            @PathVariable int messageID, @RequestBody List<Integer> attachmentIDs,
+            @AuthenticationPrincipal CustomUserDetails user) {
+        // Checks for the message, ensuring the message exists with the messageID and userID
+        if (!messagesDao.doesMessageExist(messageID, user.getUserId())) {
+            return ResponseEntity.notFound().build();
+        }
         messagesDao.editRemoveAttachments(messageID, attachmentIDs);
         return ResponseEntity.ok().build();
     }
 
-    // Delete a message with no attachments or reactions
+    // Delete a message (attachments and reactions will get deleted with CASCADE)
     @DeleteMapping("/{messageID}")
-    public ResponseEntity<Void> deleteMessage(@PathVariable int messageID) {
+    public ResponseEntity<Void> deleteMessage(
+            @PathVariable int messageID, @PathVariable int serverID,
+            @AuthenticationPrincipal CustomUserDetails user) {
+        // Checks for the message, ensuring the message exists with the messageID and userID
+        // or that the user sending the request is a mod or above
+        if (!messagesDao.doesMessageExist(messageID, user.getUserId()) &&
+                !authService.userIsModerator(user.getUserId(), serverID)) {
+            return ResponseEntity.notFound().build();
+        }
         messagesDao.deleteMessage(messageID);
         return ResponseEntity.ok().build();
     }
